requirements.txt
fastapi
sqlalchemy
pydantic
requests
geopy
uvicorn
transformers
torch
accelerate
sentencepiece
protobuf
fastapi-cache2[redis]
aiocache
redis-simple-cache